async-hello-agent/pyproject.toml
[project]
name = "async-hello-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.136.1",
    "openai-agents>=0.0.16",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.3",
]
